refactor: 공통 부분 추상화
그리고 메서드 순서를 일관되게 변경했다.
